Fix tick formatting so the numbers stay within proscribed bounds.
This required two changes. One, fix the format string passed to d3.format, I had the syntax
slightly wrong. Two, add custom logic to take ticks very close to 0 and round them to 0. This
is only necessary when the tick is near 0, because in this case the floating-point error
dominates the expression, whereas in other cases the non-error component dominates the error.
Rather than having a fixed cloesness to zero (which would break for very small domains), it
compares the cloesness to zero to the interval over which the data is spread.