Namespaced styles under ".plottable", renamed style.css to plottable.css
If a component is a root component (anchored to <svg>), it attaches the "plottable"
CSS class to its element. Descendent selectors can thus be used to apply Plottable
styles without affecting other elements on the page.

Also removed plottable.js.map, because it was confusing the Chrome debugger and
making debugging difficult.

After this change goes in the examples will need to be updated to reflect the new style sheet.

Close #148.