Add in OrdinalScale, change how BarRenderer treats x data
OrdinalScale has been added, allowing the use of category axes.
To use them you have to set the Axis' tick label formatter to
process strings (ex: function(x) { return x; }).
The scale maps inputted numbers to themselves, so the dxAccessor
in BarRenderer can be used to set the bar width with an OrdinalScale.

In conjuction with this, BarRenderer now treats data x values
as the center of each rendered bar. This change was made so that
bars would line up correctly with category ticks; I also believe it's
closer to how users expect BarRenderer to behave.