Merge Category and normal Bar Renderers, allow negative bar values
THE MERGING IS COMPLETE

The renderers were similar enough that they could be merged together. By default,
the X position of the bar is the left edge; this is the expected behavior for
generating histograms and for using rangeBands with OrdinalScales. The bar
alignment can be set via code.

Also allow arbitrary baseline values to be set. The baseline can be removed
with CSS; baseline control via methods could also be added later.

Close #250.
Close #251.