Allow removing and re-adding Components.
Changed remove() to remove a Component from the DOM, but not destroy it (as in D3)
nor to disconnect it from the IBroadcasters it was listening to.

ComponentGroup now has removeComponent() to remove a specific component from the
ComponentGroup and the DOM.

destroy() replaces the previous functionality of remove(), which disconnects a
Component from the DOM, disconnects it from all the things it was listening to, then
nulls out its element.

Attempting to anchor() a Component that's already anchor()-ed no longer throws an Error.
Instead, the Component will be re-anchor()-ed to the new DOM element.