Clean up Renderers.
First, removed the dataSelection property. In most cases, there
wasn't much point in hanging on to the selection. Exception: bar
renderers, which used the selection to support highlighting bars.

In addition, removed some unused fields and fields that were saved off,
but not used outside of the _paint() call.

Close #500.