Change Broadcaster semantics from "is-a" to "has-a"
Classes no longer extend Broadcaster. Instead, they implement the
IListenable interface which indicates that they have a broadcaster.
Listeners now attach to the Lisetnable's Broadcaster.

The goal of this change is to move to an architecture with a simpler
inheritance structure, which will hopefully better manage
complexity in the long term.

Close #534