567: Tables no longer remove null rows/columns from their interior
It used to be that if removing a component created a row and/or column
that was all nulls, the all-null row/column would be removed, potentially
moving components that were beyond the row/column left/up. The resulting
mutation to table size meant that components were no longer at the
[row, column] position the user initially specified.

This change modifies Table to remove only all-null rows/columns that have a
higher row/column index than the last non-null element (essentially, null
rows and columns will be trimmed off the outermost edge of the table).

Close #567.