Remove the null-clearing rows and associated tests.
Null rows/columns at the edge of the table are no longer trimmed.
That logic was only really relevant if a component was added at a
far-away [row, column] position and then removed. Dan says I can gloat
if the removal of this logic leads to trouble in the future. (As fun
as gloating may be, I hope I won't have to.)

Bug-specific unit tests relating to null row/columns have been removed.