Refactor so AreaPlot descends from LinePlot rather than vice versa.
Having LinePlot descend from AreaPlot was nice and made sense since you can
see a LinePlot as a special-case AreaPlot in which there is no fill, only
stroke on the line. However, this broke down once we implemented logic to
keep track of a possible baseline at the bottom of the Area, and writing special
logic in LinePlot to turn the baseline back off seemed less elegant than just having
LinePlot not implement the baseline logic at all.