Fix animations for Line and Area Plots.
Lines were animating from the top of the chart due to an oversight
(we didn't scale the zero value). They now correctly animate from 0,
or from the closest edge of the chart if zero is not in the y domain.

Area charts were updated too so that the line and area would render together.