Hover tooltips for VerticalBar Plots.
The goal was to have them work for HorizontalBar plots as well, but
Category axes don't play nice with range changes.

To use tooltips, project the desired text onto the "tooltip-text"
property. Extra space will be allocated above the bars for the tooltip.

Mousemove Interaction has been changed to a Mouse Interaction with
mouseover, mousemove, and mouseout callbacks.

The architecture still needs to be improved.