Simplify calculation, add gutter(), remove old axisTests.ts
The calculation now just takes the max over all tick lengths.
While this will truncate some things if the user zooms in, it will
work correctly for static cases.

gutter() has been added to set the amount of extra space between
the tick labels and the edge of the axis. gutter() should now
be the preferred way to give extra room for decimal places on
zoom, rather than explicitly setting the width with width().

axisTests.ts was still hanging around, so I got rid of it.