Get rid of QuantitativeScale.numTicks()
This change may be kind of controversial...

The main problem with numTicks(n) was that users were never guaranteed
to get exactly n ticks out of the Scale, so it didn't really give the
control it claimed to give. Users could only use it as an approximate
method for controlling the tick density -- but in that case, wouldn't
it make more sense to have a method that actually controlled tick density?