Fix the logic in RenderController.
"The loop was kind of pointless." -jtlan, a155f0423e2534c47e3e7db383e293b8c6b2897b

In fact, the loop was super critical, but because the logic was so convoluted,
it was difficult to tell. New logic for RenderController's flush():

Save off current list of componentsNeedingRender.
Create new list of componentsNeedingRender.
Draw all components in the saved-off list of componentsNeedingRender.
If there are still componentsNeedingRender after we finish, request another render.