Move autoranging logic into attrToExtents
Setting the Y-scale's domain based on over what's visible on X
(or vice versa) is now accomplished by filtering the data prior
to computing extents. This allowed the removal of _normalizeDatasets
and better expresses the logic behind the feature.

Now to pick better names for the methods that enable it...