Autoranging now powered by extent-setting.
Previously, some goofy filtering was done to set the Y domain as
a function of points visible on X, or vice versa.

This commit causes the dependent scale's extent to be computed with
a filter based on what points are visible on the independent scale.

Probably we should rename the methods that enable the behavior to
something more reasonable.