Move Domainer logic to Scales; delete Domainers.
Domainer logic properly belongs on the QuantitativeScales anyway, rather than
having switches inside the logic for different types of QuantitativeScales.

The Domainer unit tests all seemed to be "test Domainer in particular case
involving <some other class>". Broke up the Domainer tests and moved them to
be with the tests for that other class.

Close #2055.