Allow setting only one end of the domain on QuantitativeScale.
domainMin() sets the lower end of the domain.
domainMax() sets the upper end of the domain.

If only one is set, the other end of the scale will still be autodomained.
If both are set, it's equivalent to calling domain([min, max]).

Calling autoDomain() will clear both set values, the same way calling
autoDomain() overrides domain().