[QuantitativeScale] Compute padded domain based off of candidate domain, not current domain
Since scale() and invert() calls are used to compute the padded domain, the padded size
depends on the current domain and range, which causes problems if the domain or range
have been set to span zero width. This change temporarily sets the domain when computing
the padded domain.

Also took the opportunity to fix a test for snappingDomainEnabled to use less fragile numbers.

Close #3019.