Constrain pan/zoom to min/max domain values
Previously it was only possible to set the min/max
domain EXTENT. This was only respected by the total
zoom level.

For example, if min/max extent was set to [0, 10],
then it was possible to pan to [10, 20] or zoom to
[-5, 5].

This change adds domain VALUE constraints to both pan
and zoom to respect the absolute domain boundaries.

The implementation for zoom constraints must consider
that the zoom interaction may violate the constraints
for min, max, or both.

We handle each of these cases seperately:
min only: We adjust the "center" coordinate of the zoom.
max only: We adjust the "center" coordinate of the zoom.
min and max: We adjust both the "center" and "zoom" values.

This also supports partial constraints (i.e. only min
or max set).