Interaction: LightweightEntity Caching
Implements LightweightEntity caching in base plot in
order to speed up various chart interactions. The entity
cache is invalidated whenever underlying data changes. The
cache is constructed lazily on the first new call to `entities()`
from the time of invalidate (or initialization).

Entity positioning is stored as Points in the Data space, rather
than the screen space in order to avoid having to rebuild the
cache on each pan/zoom interaction. Comparison is done on chart
and cursor positions which are all converted to data space first.

-- notable changes

* Base plot now implements entity caching.
* `LightweightPlotEntity` is now exported publicly.
* Adds `TransformableAccessorScaleBinding`.
* Moves `TransformableScale` from PanZoomInteraction
  to `Plottable.TransformableScale` namespace.
* Quantitative Scales accessors x(), y(), x1(), y1() now
  return an instance of `TransformableScale` instead of
  simply `Scale`
* Adds `utils/entityIndex.ts`, which defines
  and implements a simple entity storage API
* New public method `bounds` on component returns
  the {Bounds} of the component
* Adds `invertTransformation` method to `TransformableScale`
  which maps pixel coordinates to Transformation Space