Add `stackingOrder` property to stacked bar and stacked area charts
By default, stacked plots are "bottomup", meaning for positive data, the
first series will be placed at the bottom of the scale and subsequent data
series will by stacked on top. This can be reveresed by setting stacking order
to "topdown".

This makes it easier to have a legend that matches the order of the stack since
legends will display the first series at the top of the list.