Routable Quicktests
The "overlaying" quicktests can't be individual linked to.

Moreover livereload resets your view so you have to re-select
the tests you were working on every time.

This commit adds a simple list of tests that can be linked
to with hashroutes so livereload updates will maintain
your place and you can work with tests one at a time.