Fix pan-zoom interaction for reversed domains and ranges
Domains with max < min (like [-10, 0]) work just fine
with the pan-zoom interaction, but we had explicit tests
disabling this. The commit removes those conditions and
removes the unit tests that confirmed this.

However, there was an issue with reversed ranges since
all pan-zoom constraint calculations happen in range space.
This commit includes fixes for those calculations and
adds a quicktest that has every combination of x/y reversing
in scatter plots.