pull in d3 using d3/build/d3.node
d3 specifies a "browser": "build/d3.js" in its package.json which means
bundlers will grab this pre-built d3, which doesn't play well with
d3-selection-multi. It'd also make bundlers bundle all the modules
twice: once in the built file and again as a micro module.

To solve this, we grab the node file directly