Histogram Bar Plots
Add barAlignment getter/setter for bar plots to
change how bars are aligned to the x accessor
coordinate.

Add barEnd accessor to bar plots which is used
to calculate individual bar widths as opposed
to using the same width for all bars. When
set the singular _barPixelWidth value is
not recomputed.

Add tests and update grunt-blanket-mocha so
that phantom runs on macOS Sierra