better bar plot terminology; stacked datum wip
enable bar labels on histogram example

use the term bar "thickness" instead of bar "width" where appropriate.
width is specifically associated with the x direction but it was being
used to represent the thickness (which is meaningful in both x and y).

Also consolidate the terms "primary" and "secondary" into bar "length"
and bar "position". Add methods to get the "length" binding and
"position" bindings.

add a reference back to the original datum/index/dataset for each
stacked datum