Fix drawLabels with barAlignment (#3370)
* better bar plot terminology; stacked datum wip

enable bar labels on histogram example

use the term bar "thickness" instead of bar "width" where appropriate.
width is specifically associated with the x direction but it was being
used to represent the thickness (which is meaningful in both x and y).

Also consolidate the terms "primary" and "secondary" into bar "length"
and bar "position". Add methods to get the "length" binding and
"position" bindings.

add a reference back to the original datum/index/dataset for each
stacked datum

* bar thickness variable

use a variable "thickness" to denote which places in the code
semantically mean thickness vs. width

* always pass the center of the bar to computeLabelTopLeft

more width -> thickness renames

* allow drawing out-of-bounds extrema labels

* fix linting

* delete extraneous tests; rename try->maybe