memoize compute extents (#3385)
* install is-plain-object dep

* move memoize and sign; add memThunk; memoize computeExtents

memThunk is a simple wrapper around memoize that, similar to reselect's
selectors, lets you define a "thunk" functions (functions that
take "no input") that represents computed state based off of internal
state

We use this to memoize computeExtents in plot.ts and delete
the manual _updateExtents calls. This improves
readability, maintainability and performance.

This introduced a bug - updating extents is sometimes overridden by
subclasses. This should be addressed

* fix passing scale as postScale accessor in tests

* memoize stack extents in stacked bar

* memoize stacking result for stacked area

* slice the right inputs for memoization

* fix tests

* fix linting

* delete commented out error

* split into multiple files

* add tests for memoize, signature api, and memthunk

* fix linting

* point yarn.lock to public registry

for some reason you have to remove the registry key from ~/.npmrc; i
could not override in .yarnrc

* comment memoize; better function names