Fix deferred rendering (#3415)
The primary cause of incorrect deferred rendering results was that
the deferred renderer did not know when a component was rendered
except when it triggered the render. Occassionally, direct calls
to render would occur and the deferred renderer did not update
the css transform. This PR refactors the code so that the
component's own render method will reset the deferred renderer to
guarantee that it resets the css transform.

- Add `renderLowPriority` method to all components to make typing easier
- Extract transform method from constructor for legibility
- Switch to transformable scales to support category axis deferred render